+++ client/src/components/Calculator/CreditFormStepper.js
@@ -7,14 +7,11 @@ import Button from '@material-ui/core/Button';
 import Typography from '@material-ui/core/Typography';
 import Box from '@material-ui/core/Box';
 
-
-import FormControl from '@material-ui/core/FormControl';
-
 import CreditType from './CreditType'
 /*import CreditAffiliation from './CreditAffiliation'*/
 import CreditAmount from './CreditAmount'
 import CreditInsurance from './CreditInsurance'
-import { Card, CardContent, CardActionArea } from '@material-ui/core';
+import { Card, CardContent } from '@material-ui/core';
 
 
 const getSteps = function() {
