+++ client/src/components/Unused/OfferRowTable.js
