+++ client/src/components/landingPage/SimpleMenu.js
